Configured frontend proxy to always have the /api prefix to avoid token expiration issue with /auth and /admin endpoints
