+++ src/routes/admin/jobs/+page.svelte
@@ -40,7 +40,7 @@
 		/** @type {import('$lib/types').ApplyWorkflow[]} */
 		const updatedJobs = [];
 		for (const job of jobsToCheck) {
-			const url = new URL('/admin/job', window.location.origin);
+			const url = new URL('/api/admin/job', window.location.origin);
 			url.searchParams.append('id', job.id.toString());
 			const response = await fetch(url);
 			if (response.ok) {
@@ -60,7 +60,7 @@
 			if (searchErrorAlert) {
 				searchErrorAlert.hide();
 			}
-			const url = new URL('/admin/job', window.location.origin);
+			const url = new URL('/api/admin/job', window.location.origin);
 			if (status) {
 				url.searchParams.append('status', status);
 			}
